backend/requirements.txt
flask
flask-cors
sentence-transformers
openai
python-dotenv